removing hash map for client_id

+++ shared/src/config.rs
@@ -1,6 +1,6 @@
 pub const KEY_SERVER_SUFFIX: &str = ".t.craftip.net";
 pub const SERVER_PORT: u16 = 25565;
-pub const MAXIMUM_CLIENTS: u16 = 255;
+pub const MAXIMUM_CLIENTS: usize = 255;
 pub const PROTOCOL_VERSION: u16 = 1;
 pub const TIMEOUT_IN_SEC: u64 = 20;
